drivers/devfreq: adreno_tz: Fix GPU target frequency calculation for high refresh rates
It seems that the black-box GPU frequency selection algorithm is unaware
of high refresh rates, and assumes a refresh rate of 60 Hz for its
target frequencies. This results in frequencies that are too low being
selected when a higher refresh rate, such as 90 Hz, is used.

To remedy this, multiply the GPU's busy time by a ratio of the current
refresh rate to 60 Hz. This makes msm-adreno-tz select the right
frequencies needed to avoid stuttering.

Change-Id: Ia1836630a6a0fae238cc4cd7e159aba6a431518d
Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
[ hridaya: Adapt to k4.19 changes ]
Signed-off-by: hridaya <hridaya@pixelos.net>