ARM64/configs: gauguin: Don't randomize module region over a 4 GB range
Randomizing a module is for security purpose to deter exploit
attempts relying on knowledge of the location of kernel internals.

But randomizing the location of module region inside a 4 GB window
to prevent core kernel data leakage is expensive. We are better off
with a limited range. Randomizing module region is expensive anyway.

Change-Id: Iad6fdfa3fe902a589ab86c77de40512bf5878baa
Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>
Signed-off-by: hridaya <hridaya@pixelos.net>