zram: Kconfig: Use selections for CRYPTO_{LZO, ZSTD, LZ4}
If we use "depends on" for the default zram compression algorithm,
all of the crypto_{lz4, lzo, zstd} are needed to be enabled.
Otherwise, zram will miss dependencies and disable itself.
To fix that, use "select" instead of "depends on".

As we are using selections now, remove the dependencies at ZRAM
for crypto_{lz4, lzo, zstd}.

Fixes: 4e565febb ("BACKPORT: zram: break the strict dependency from lzo")
Acked-by: Divyanshu-Modi <divyan.m05@gmail.com>
Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>
Signed-off-by: hridaya <hridaya@pixelos.net>