usb: gadget: f_fs: Remove unused variables
With commit [1], our original intention was to just remove
ipc log string and resolve the uninitialized variable warn
for our benefits. It seems commit [1] made that variable's
use case unused and also the ones who are dependent on it
and the similars.

This removes all of the unused variables detected by
GCC after commit [1].

[1]: 5b837a8e1380 ("usb: gadget: f_fs: Make ipc log strings no-ops")

Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>
Signed-off-by: hridaya <hridaya@pixelos.net>
Change-Id: I9b60057260e1d7583ffbc40d5cdb2371a67fd040