simple_lmk: Default to stock vmpressure trigger threshold
Vmpressure trigger threshold of >= 80 causes apps to be killed very
frequently and thus apps are failing to be consistently kept in
the background on a 6GB ram device. It is to be noted that vmpressure
is the only trigger for slmk to kill app and reclaim memory based on a
specific configuration of minfree and its timeout.
Therefore, switch back to the stock threshold.

Change-Id: Ie9dd06aa801427be3d69bf1d7bd16281cbce2d5f
Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>
Signed-off-by: hridaya <hridaya@pixelos.net>